refactor: move app code to cmd/waku
